requirements.txt
numpy
matplotlib
gymnasium
stable-baselines3
roboticstoolbox-python
scipy
pygame
